Move PopupBridge module to com.braintree.payments.api package.
