+++ PopupBridge/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.braintreepayments.popupbridge"/>
+<manifest package="com.braintreepayments.api.popupbridge"/>
