+++ PopupBridge/src/main/java/com/braintreepayments/api/PopupBridgeMessageListener.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.popupbridge;
+package com.braintreepayments.api;
 
 public interface PopupBridgeMessageListener {
     void onMessageReceived(String messageName, String data);
