+++ PopupBridgeExample/src/androidTest/java/com.braintreepayments.popupbridge.example.test/PopupBridgeTest.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.popupbridge.example.test;
+package com.braintreepayments.api.example.test;
 
 import android.os.Build;
