+++ PopupBridgeExample/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.braintreepayments.popupbridge.example">
+    package="com.braintreepayments.api.example">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -10,16 +10,16 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light">
 
-        <activity android:name=".PopupActivity"/>
+        <activity android:name="com.braintreepayments.api.example.PopupActivity"/>
 
-        <activity android:name=".MainActivity">
+        <activity android:name="com.braintreepayments.api.example.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
-        <activity android:name="com.braintreepayments.popupbridge.PopupBridgeActivity"
+        <activity android:name="com.braintreepayments.api.PopupBridgeActivity"
             android:launchMode="singleTask">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
